requirements.txt
aiogram==3.6.0
gspread_asyncio==2.0.0
google-auth==2.29.0
python-dotenv==1.0.1
pytest==8.2.0
